+++ NEWS.md
@@ -1,3 +1,8 @@
+breakDown 0.2.0
+----------------------------------------------------------------
+* `break_down` function identifies inteactions 
+* `break_down` function has complexity O(2p) for models without interactions, much faster than the old version
+
 breakDown 0.1.6
 ----------------------------------------------------------------
 * `broken.default` has now the `keep_distributions` arguments. If `TRUE` then the whole distribution of conditional residuals is remebered and avaliable for plotting   [#17](https://github.com/pbiecek/breakDown/issues/17)